requirements.txt
absl-py==0.8.1
astor==0.8.0
cffi==1.13.2
Click==7.0
cycler==0.10.0
decorator==4.4.1
editdistance==0.5.3
Flask==1.1.1
fpdf==1.7.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
nltk==3.4.5
numpy==1.17.2
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.1
pdf2image==1.9.0
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyvips==2.1.8
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
Werkzeug==0.16.0
wrapt==1.11.2
